+++ R/app_annotate.R
@@ -1645,7 +1645,7 @@ annotate_server <- function(id) {
       trigger("annotations_modal")
     })
 
-    annotations_details_server(ns("annotations"), rv)
+    annotations_details_server(ns("annotations"), rv, table_id = ns("table"))
 
     # CSV Export ----
     .export_cols_drop <- c("output", "view", "poor_blast_ref", "warnings_details", "blast_accession_auto")